+++ libp2p/src/lib.rs
@@ -137,6 +137,11 @@ pub use libp2p_websocket_websys as websocket_websys;
 #[cfg_attr(docsrs, doc(cfg(feature = "webtransport-websys")))]
 #[doc(inline)]
 pub use libp2p_webtransport_websys as webtransport_websys;
+#[cfg(feature = "webtransport")]
+#[cfg(not(target_arch = "wasm32"))]
+#[cfg_attr(docsrs, doc(cfg(feature = "webtransport")))]
+#[doc(inline)]
+pub use libp2p_webtransport as webtransport;
 #[cfg(feature = "yamux")]
 #[doc(inline)]
 pub use libp2p_yamux as yamux;